refactor: LottoLineGenerator 의 패키지를 application 으로 이동했다.
