+++ src/main/kotlin/lotto/application/LottoLineGenerator.kt
@@ -1,4 +1,6 @@
-package lotto.domain
+package lotto.application
+
+import lotto.domain.LottoLine
 
 fun interface LottoLineGenerator {
     fun generate(): LottoLine
