+++ src/test/kotlin/lotto/application/RandomLottoLineGeneratorTest.kt
@@ -1,7 +1,8 @@
-package lotto.domain
+package lotto.application
 
 import io.kotest.inspectors.shouldForAll
 import io.kotest.matchers.shouldBe
+import lotto.domain.LottoNumber
 import org.junit.jupiter.api.Test
 
 @Suppress("NonAsciiCharacters")
